BURNING THE CHURCH

+++ src/resources/controllers/board/boardController.ts
@@ -1,12 +1,13 @@
 import Controller from '@/utils/interfaces/controllerInterface';
 import { Router, Request, Response } from 'express';
-import auth from '@/middleware/auth.middleware';
-import userModel from '@/resources/models/userModel';
-import boardModel from '@/resources/models/boardModel';
 import { boolean, string } from 'zod';
 import z from 'zod';
 import cardModel from '@/resources/models/cardModel';
 import mongoose from 'mongoose';
+import auth from '@/middleware/auth.middleware';
+import userModel from '@/resources/models/userModel';
+import boardModel from '@/resources/models/boardModel';
+import raiaModel from '@/resources/models/raiaModel';
 
 class BoardController implements Controller {
     public path = '/board';
@@ -24,7 +25,7 @@ class BoardController implements Controller {
         /**
          * Listar boards do usuário, o parametro filtra com all, fav, shared
          */
-        this.router.get(`${this.path}/:id`, auth, this.listMyBoard);
+        this.router.get(`${this.path}/:filter`, auth, this.listMyBoard);
         /**
          * Listar cards de um determinado board
          */
@@ -35,83 +36,96 @@ class BoardController implements Controller {
         this.router.delete(`${this.path}/:id`, auth, this.deleteBoard);
     }
 
-    private async createNewBoard(req: Request, res: Response): Promise<void> {
+    private async createNewBoard(req: Request, res: Response): Promise<any> {
+        const session = await mongoose.startSession();
+        session.startTransaction();
         try {
             const user = await userModel.findOne({ _id: req.userId });
 
             if (!user) {
-                res.status(404).json({ message: 'User not found' });
-            } else {
-                const newBoard = z.object({
-                    nome: string(),
-                    cor: string(),
-                    icon: string(),
-                    favorito: boolean(),
-                });
-
-                const { nome, cor, icon, favorito } = newBoard.parse(req.body);
-
-                const board = await boardModel.findOne({
-                    nome: nome,
-                    owner: user._id,
-                });
-
-                if (!board) {
-                    const data = await boardModel.create({
-                        nome: nome,
-                        cor: cor,
-                        icon: icon,
-                        favorito: favorito,
-                        owner: user._id,
-                        followers: [user._id],
-                    });
-
-                    res.status(201).json({ data });
-                } else {
-                    res.status(401).json({ message: 'Board already exists' });
-                }
+                throw new Error('User not found');
             }
-        } catch (error) {
-            res.status(401).json(error);
+
+            const newBoard = z.object({
+                nome: string(),
+                cor: string(),
+                icon: string(),
+                favorito: boolean(),
+            });
+
+            const { nome, cor, icon, favorito } = newBoard.parse(req.body);
+
+            const board = await boardModel.findOne({
+                nome: nome,
+                owner: user._id,
+            });
+
+            if (!board) {
+                const data = await boardModel.create({
+                    nome: nome,
+                    cor: cor,
+                    icon: icon,
+                    favorito: favorito,
+                    owner: user._id,
+                    followers: [user._id],
+                });
+                await session.commitTransaction();
+                return res.status(201).json({ data });
+            } else {
+                return new Error('Board already exists');
+            }
+        } catch (error: any) {
+            await session.abortTransaction();
+            if (error.message === 'User not found') {
+                return res.status(404).json({ message: 'User not found' });
+            }
+            if (error.message === 'Board already exists') {
+                return res
+                    .status(401)
+                    .json({ message: 'Board already exists' });
+            }
+
+            return res.status(401).json(error);
+        } finally {
+            session.endSession();
         }
     }
-    private async listMyBoard(req: Request, res: Response): Promise<void> {
+    private async listMyBoard(req: Request, res: Response): Promise<any> {
         try {
             const user = await userModel.findOne({ _id: req.userId });
             if (!user) {
-                res.status(404).json({ message: 'User not found' });
-            } else {
-                if (!req.params.id || req.params.id === 'all') {
-                    const boards = await boardModel.find({
-                        owner: user._id,
-                    });
-                    res.status(200).json({ boards });
-                } else {
-                    if (req.params.id === 'fav') {
-                        const boards = await boardModel.find({
-                            owner: user._id,
-                            favorito: true,
-                        });
-                        res.status(200).json({ boards });
-                    } else {
-                        if (req.params.id === 'shared') {
-                            const boards = await boardModel.find({
-                                $or: [
-                                    { owner: user._id },
-                                    { followers: user._id },
-                                ],
-                            });
-                            res.status(200).json({ boards });
-                        } else {
-                            res.status(400).json({
-                                message:
-                                    'Unknown param; It can only be: all, shared or fav',
-                            });
-                        }
-                    }
-                }
+                throw new Error('User not found');
             }
-        } catch (error) {
+            if (
+                !req.params.filter ||
+                req.params.filter === 'all' ||
+                req.params.filter === ''
+            ) {
+                const boards = await boardModel.find({
+                    owner: user._id,
+                });
+                return res.status(200).json({ boards });
+            }
+            if (req.params.filter === 'fav') {
+                const boards = await boardModel.find({
+                    owner: user._id,
+                    favorito: true,
+                });
+                return res.status(200).json({ boards });
+            }
+            if (req.params.filter === 'shared') {
+                const boards = await boardModel.find({
+                    $or: [{ owner: user._id }, { followers: user._id }],
+                });
+                return res.status(200).json({ boards });
+            }
+            return res.status(400).json({
+                message: 'Unknown param; It can only be: all, shared or fav',
+            });
+        } catch (error: any) {
+            if (error.message === 'User not found') {
+                return res.status(404).json({ message: 'User not found' });
+            }
             res.status(401).json(error);
         }
     }
@@ -122,43 +136,64 @@ class BoardController implements Controller {
             const user = await userModel.findOne({ _id: req.userId });
 
             if (!user) {
-                return res.status(404).json({ message: 'User not found' });
+                throw new Error('User not found');
             }
 
-            if (!req.params.id) {
-                return res.status(400).json({ message: 'Param not found' });
+            if (!req.params.id || req.params.id === '') {
+                throw new Error('Param not found');
             }
+
+            const raia = await raiaModel.findOne({ board: req.params.id });
 
             const cards = await cardModel.deleteMany({
-                board: (req.params.e as any).algumacoisa,
+                board: { $in: raia?.cards },
             });
-            const board = await boardModel.deleteOne({ _id: req.params.e });
+
+            const board = await boardModel.deleteOne({ _id: req.params.id });
+
             await session.commitTransaction();
-            return res.status(200).json({ message: 'ok ok ' });
-        } catch (error) {
+
+            return res.status(200).json({ message: 'Success on delete!' });
+        } catch (error: any) {
             await session.abortTransaction();
+
+            if (error.message === 'User not found') {
+                return res.status(401).json({ message: 'User not found' });
+            }
+            if (error.message === 'Param not found') {
+                return res.status(401).json({ message: 'Param not found' });
+            }
+
             return res.status(401).json({ error: 'Something went wrong' });
         } finally {
             session.endSession();
         }
     }
-    private async getThisBoard(req: Request, res: Response): Promise<void> {
+    private async getThisBoard(req: Request, res: Response): Promise<any> {
         try {
             const user = await userModel.findOne({ _id: req.userId });
             if (!user) {
-                res.status(404).json({ message: 'User not found' });
-            } else {
-                if (!req.params.id) {
-                    res.status(400).json({ message: 'Param not found' });
-                } else {
-                    const cards = await cardModel.find({
-                        board: req.params.id,
-                        users: user._id,
-                    });
-                    res.status(200).json({ cards });
-                }
+                throw new Error('User not found');
             }
-        } catch (error) {
+            if (!req.params.id) {
+                throw new Error('Param not found');
+            }
+
+            const raia = await raiaModel
+                .find({
+                    board: req.params.id,
+                    users: user._id,
+                })
+                .populate('cards');
+
+            res.status(200).json({ raia });
+        } catch (error: any) {
+            if ('User not found') {
+                return res.status(404).json({ message: 'User not found' });
+            }
+            if ('Param not found') {
+                return res.status(404).json({ message: 'Param not found' });
+            }
             res.status(401).json(error);
         }
     }
